Tests passing + integration tests building

+++ src/Stratis.SmartContracts.IntegrationTests/ContractExecutionFailureTests.cs
@@ -116,7 +116,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.CreateCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.CreateCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Null(receipt.To);
@@ -160,7 +160,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.CreateCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.CreateCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Null(receipt.To);
@@ -306,7 +306,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.BaseCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.BaseCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Equal(StateTransitionErrors.NoCode, receipt.Error);
@@ -352,7 +352,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.BaseCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.BaseCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Equal(ContractInvocationErrors.MethodDoesNotExist, receipt.Error);
@@ -405,7 +405,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.BaseCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.BaseCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Equal(ContractInvocationErrors.MethodDoesNotExist, receipt.Error);
@@ -457,7 +457,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.CreateCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.CreateCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.StartsWith(ContractInvocationErrors.MethodDoesNotExist, receipt.Error); // The error for constructor not found vs method does not exist could be different in future.
@@ -514,7 +514,7 @@ namespace Stratis.SmartContracts.IntegrationTests
             Assert.Equal(response.TransactionId.ToString(), receipt.TransactionHash);
             Assert.Empty(receipt.Logs);
             Assert.False(receipt.Success);
-            Assert.Equal(GasPriceList.CreateCost, (long) receipt.GasUsed);
+            Assert.Equal(GasPriceList.CreateCost, receipt.GasUsed);
             Assert.Null(receipt.NewContractAddress);
             Assert.Equal(this.node1.MinerAddress.Address, receipt.From);
             Assert.Null(receipt.To);
